Added a form to the home page.
Note: It is still a static page.

It posts the form.

A @Controller annotated class will handle the POST.  The form data is automatically mapped to an object using the names in the form and the names of the attributes to match.

The name is output to the console (must change this to logging) and the user is redirected back to the home page.  Note, how this is done with a "redirect:".