Added a singleton class to play the role of a CharityService.
We don't want this type of responsibility in the controller.

Later on, we will replace this singleton with a Spring Bean.

Note the use of an interface.  This will allow the controller to easily replace the singleton.

The results of the call to the CharityService are added to the model.  This "model" is part of the MVC (model-view-controller) pattern implementation in Spring.  It is the part of the domain that the next web screen requires to render itself.  The model consists of a set of key-value pairs that we pass back to the view.  We return the name of the view.  By convention, this is just the name of the HTML file in the templates folder.

The template uses the Thymeleaf template engine.  Thymeleaf can access the fields by key.  An example of iteration is shown as well as some in-page display logic.