Added a REST api to obtain charity data.
Added a dust.js template to inject charity data into the donor page.

This keeps the data on the session and/or on the URL to a minimum as the charity data can be obtained on demand.

Is this good or bad?  It increases network traffic and server load, but the charity data could be cached as it rarely changes?  It reduces the session size which, in production, would reduce network traffic as the session is replicated across hosts.

Also added a REST test.