Added two new charities.
We've now discovered that charities can share a registration number, but operate separate accounts.

This is how the Cats Protection League operate.

This means that the registration id can now longer act as a natural primary key.

We need to consider our options and how this will change our interfaces.

We can introduce an artificial key and/or we could introduce a furl that would need to be checked at charity creation time for uniqueness.

We will do both.