simplified API down to a search function and a charity by id function.
Big refactoring done here.  The tests helped make sure things were back, but the tests need to be strengthened with more assertions so that potential issues can be found quicker.

For example, the tests should assert on the model as well as on the content because the failure can help diagnose where the issue lies.