This version uses exceptions.
The service throws an exception when it can't find a charity.  Since it is a RunTime exception, it doesn't have to be caught in the controller.

The controller advice will still handle it.

Which is better?

Given Java 8 provides the Optional<> feature, should we use this?  The only reason for not using it, is that the tests don't work because the model is an Optional and the matchers can't access the properties.  Maybe we should look for a way to get past this and maintain a policy of Exceptions for exceptional conditions only?