Fixed Javascript issues.
Solution was to put everything at bottom of page, which seems odd.

Now moving onto the saving of a donation.  We're basing the objects on the DB design, so a donation is made by a donor who has an address.

This is useful because a charity might want to know all their donors - but there is an issue here.  What if the donor moves after they make a donation?  Why would they inform this donation system?