A lot has happened in this commit.
Domain objects:  We now have Donation, Donor and Address classes.  Note: A donation refers to a donor and an donor to an address.  The classes do not contain foreign key attributes.  Rather they contain object references.  But our database is MySQL and that requires foreign keys.

We have added a Donation repository ONLY.  There is no Donor or Address repo.  From the point of view of creating a Donation, it makes no sense to save the parts separately.  In essence, the Donation is the root of an aggregate which is a design term used in Domain-Driven Design (DDD) by Eric Evans.  DDD is well beyond the scope of this module as it could be the basis of a module in its own right.

We have a Donation Service that offers a method to create a donation.  This means that the controller has to form a complete Donation object and pass it to the Service.  We don't want the Service to know about the various forms that are created as different clients could create different UIs.  So, the controller has to marshall the data into the Donation object.

This is then passed through to a Repository.  In DDD, we may by-pass the service and go straight to the Repository because the Service isn't doing anything extra here.

Pay attention to the repository JDBC code.

We use a Prepared Statement because this offers better protection against SQL injection attacks.  A prepared statement is a piece of SQL with parameters indicated by "?".  Once created, the parameters can be provided before the statement is executed.

Note also the use of the generated key code.  When we create an Address in the database, MySQL will automatically create the id of the Address.  The SQL to create the donor requires this because it has to include the foreign key in the donor insert.

This is then repeated for the donation.

This mapping of a donation object into 3 SQL inserts is hidden from the client of the service/repository.

You can see that this is quite complex code (though it follows a standard pattern).

We can ask a framework to handle this mapping for us - and we'll look at that shortly.

We are now in need of some tests!