With this commit, we've introduced the ideas of test scripts and profiles.
In src/main, we've introduced H2 and MySQL profiles.

We can switch databases by telling Spring at run-time to run in a certain profile.

However, this highlights an issue.  SQL isn't the same in every database.  The date parser is different.

We want to use H2 for testing because it is faster, so we write a version of the data load script specifically for that.

In the real-world, we would maintain a parallel version of the MySQL code, or look at a tool like DBUnit.

spring.profiles.active=h2 tells Spring that the H2 profile should be used.

In test, we can have special properties and SQL files.  When we test, we automatically set up and populate a H2 file.

Note: this process took a long time to work out.  It was an example of how frameworks have opinions and developers have to read a lot of documentation to work out how to follow those opinions through to working code.

What we want to do next is add some real E2E tests.