Changed persistence to JPA.
Most of the SQL is now written by Spring (using Hibernate as an implementation of JPA).

I use @Query to support native queries where the SQL can't be generated.

Only change to the repo interface was to make findOne into findById on charity.  This was an initial mistake.  I should have started with findById.

Look carefully at build.gradle and the application*.properties files to see how they are configured to support build, bootrun and MySQL running.