Documented the other options for the Donation Repository.
This is a good example of developing with frameworks.  Some things may be undocumented or poorly documented - and you may need to dig under the covers to find out what is happening.

You will often be left with a set of choices.

You may also have a dilemma in that the interface you wish to offer clashes with the framework.

In this application, I want a DDD-style use of Repository so I've created a DonationRepository interface.  I've been aware of Spring Boot's naming standards for @Repository beans.

However, in this case, my simple search API isn't easily achieveable with Spring Data JPA.

I've decided to do it with a like clause everywhere because, in hindsight, partial matching of all fields is probably a good thing to have.